+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Added `unknown` key to `languages` which lists file extensions that cannot be matched with a given language.
+
 ## 1.3.1
 - Changed the delimiter for ignored paths in the CLI `--ignored` argument from `;` to `:`/`;`/`|` as a semicolon is actually valid on Windows.